Documentation:
- Vulkan Validation layers documentation has been updated
- build requirements have been updated

Project:
- Android Studio has been updated to Arctic Fox | 2020.3.1 Patch 4
- Android Studio Gradle Plugin has been updated to 7.0.4
- Gradle plugin has been updated to 7.3.1-all
- Vulkan Validation layers has been updated to 0965e77eb083b8b4becf01cd8aa39284e56e6338
- DirectX Shader Compiler has been updated 1.6.2112.10024
- SPIR-V shader blobs have been recompiled (no changes)

third-party:
- Vulkan Validation layers has been recompiled
- CMakeLists.txt for Vulkan Validation layers has been updated

GXCommon:
- unit axis requirement was added to GXVec3::Project
- clang-tidy fixes
- GXMat4 API has been extended
- code style

Android:
- unused resources have been removed
- manifest fix

android-vulkan:
- GJK refactoring
- assign name for texture 2D
- debug code has been removed
- GJK refactoring
- velocity solve iterations have been changed to 5 instead of 7
- ray cast query returns rigid body object
- better Vulkan extension checking
- vkGetPhysicalDeviceFeatures2 was added
- shader int8 requirement
- VK_KHR_16bit_storage requirement has been removed
- no VK_KHR_16BIT_STORAGE_EXTENSION_NAME
- no float16_t interpolators
- Adreno 660 support
- mesh geometry loading fix
- gamepad dead lock scenario has been resolved
- GJK raycast
- sweep test
- RigidBody context support
- clang-tidy fixes
- code style

pbr:
- camera movement speed
- Submit as part of public StaticMeshComponent API
- material lookup fix
- opaque material refactroing
- point light component API altering
- static mesh component material API
- unorm material color refactoring
- palette color refactoring
- material palette optimal rendering
- emission color support
- Static Mesh component refactoring
- stop camera after release input operation
- clang-tidy fixes
- code style

ray_casting:
- new subproject

sweep_testing:
- new subproject

+++ app/src/main/cpp/sources/epa.cpp
@@ -68,17 +68,17 @@ uint16_t EPA::GetSteps () const noexcept
 
 uint16_t EPA::GetEdgeCount () const noexcept
 {
-    return _edges.size ();
+    return static_cast<uint16_t> ( _edges.size () );
 }
 
 uint16_t EPA::GetFaceCount () const noexcept
 {
-    return _faces.size ();
+    return static_cast<uint16_t> ( _faces.size () );
 }
 
 uint16_t EPA::GetVertexCount () const noexcept
 {
-    return _vertices.size ();
+    return static_cast<uint16_t> ( _vertices.size () );
 }
 
 void EPA::Reset () noexcept
